requirements.txt
Flask
transformers
torch
mysql-connector-python
python-dotenv